+++ lib/hcloud/resources/load_balancer_type.rb
@@ -33,6 +33,8 @@ module HCloud
 
     attribute :prices, :price, array: true, default: -> { [] }
 
+    attribute :deprecation, :deprecation
+
     attribute :deprecated, :datetime, deprecated: true
   end
 end