+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add `deprecation` attribute to `HCloud::LoadBalancerType`
 - Add `MustBeUnassigned` error class
+- Add `DeprecatedApiEndpoint` error class
 
 ### Changed